Обновление механик пушечных снарядов
- Пушечный снаряд взрывается при не пробитии стены техники
- Мощность взрыва зависит от калибра снаряда

+++ code/modules/projectiles/projectile.dm
@@ -516,16 +516,19 @@
 	if (T != firer_loc)
 		for (var/obj/structure/vehicleparts/frame/F in T.contents)
 			var/penloc = F.get_wall_name(direction)
+			F.bullet_act(src,penloc)
 			if (!F.CheckPen(src,penloc))
 				passthrough = FALSE
 				visible_message("<span class = 'warning'>Снаряд не пробивает [penloc] стену!</span>")
-				F.bullet_act(src,penloc)
 				bumped = TRUE
-				loc = null
-				qdel(src)
+				if (istype(src, /obj/item/projectile/shell))
+					var/obj/item/projectile/shell/S = src
+					S.initiate(permutated[permutated.len])
+				else
+					loc = null
+					qdel(src)
 				return FALSE
 			else
-				F.bullet_act(src,penloc)
 				passthrough = TRUE
 				forceMove(T)
 				permutated += T
@@ -533,6 +536,7 @@
 
 	if (!is_trench && launch_from_trench && !overcoming_trench)
 		overcoming_trench = TRUE
+
 	if (T.density)
 		passthrough = FALSE
 	else
@@ -681,6 +685,12 @@
 			passthrough = TRUE
 			passthrough_message = "<span class = 'warning'>Пуля  пробивает насквозь [T]!</span>"
 		--penetrating
+
+	if (istype(src, /obj/item/projectile/shell))
+		if (loc == trajectory.target)
+			var/obj/item/projectile/shell/S = src
+			S.initiate(loc)
+			return FALSE
 
 	//the bullet passes through the turf
 	if (passthrough)
